+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### Improvements
 
+- [#95][] Make `Listen::Listener` capable of listening to multiple directories. ([@rymai][])
+- [#95][] Deprecate `Listen::MultiListener`. ([@rymai][])
 - Add `Listen::Adapter#pause`, `Listen::Adapter#unpause` and `Listen::Adapter#paused?`. ([@rymai][])
 - Refactor `Listen::DirectoryRecord` internals. ([@rymai][])
 - Refactor `Listen::DependencyManager` internals. ([@rymai][])
@@ -227,6 +229,7 @@
 [#85]: https://github.com/guard/listen/issues/85
 [#88]: https://github.com/guard/listen/issues/88
 [#93]: https://github.com/guard/listen/issues/93
+[#95]: https://github.com/guard/listen/issues/95
 [@Maher4Ever]: https://github.com/Maher4Ever
 [@akerbos]: https://github.com/akerbos
 [@antifuchs]: https://github.com/antifuchs