+++ CHANGELOG.md
@@ -12,11 +12,11 @@
 
 ### Improvements
 
-- [#95][] Make `Listen::Listener` capable of listening to multiple directories, deprecate `Listen::MultiListener`. ([@rymai][])
 - New method: `Listen.to!` which blocks the current thread. ([@rymai][])
 - New method: `Listen::Listener#start!` to start the listener and block the current thread. ([@rymai][])
+- [#95][] Make `Listen::Listener` capable of listening to multiple directories, deprecates `Listen::MultiListener`. ([@rymai][])
 - [#85][] Compute the SHA1 sum only for regular files. ([@antifuchs][])
-- Add `Listen::Adapter#pause`, `Listen::Adapter#unpause` and `Listen::Adapter#paused?`. ([@rymai][])
+- New methods: `Listen::Adapter#pause`, `Listen::Adapter#unpause` and `Listen::Adapter#paused?`. ([@rymai][])
 - Refactor `Listen::DirectoryRecord` internals. ([@rymai][])
 - Refactor `Listen::DependencyManager` internals. ([@rymai][])